Move out vg translate
